pyproject.toml
[project]
name = "crypto-forecasting"
version = "0.1.0"
description = "A package for forecasting cryptocurrency prices."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ccxt==4.5.32",
    "fh-plotly==0.2.4",
    "httpx==0.28.1",
    "loguru==0.7.3",
    "monsterui==1.0.36",
    "numpy==2.4.1",
    "pandas==2.3.3",
    "plotly==6.5.1",
    "pydantic==2.12.5",
    "pydantic-settings==2.12.0",
    "python-dotenv==1.2.1",
    "python-fasthtml==0.12.39",
    "scikit-learn==1.8.0",
    "sqlalchemy==2.0.45",
]
